requirements.txt
pandas>=1.5.0
openpyxl>=3.0.10
mysql-connector-python>=8.0.30
python-dotenv>=0.21.0
sqlalchemy>=1.4.0
